Refactor pipeline to put steps in their own files.
