+++ src/decontext/logging.py
@@ -1,2 +1,6 @@
 def warn(message: str):
     print(f"[WARNING] {message}")
+
+
+def info(message: str):
+    print(f"[INFO] {message}")
